chore: new desktop app
