+++ .idea/artifacts/GameOutlet_jvm_1_0_SNAPSHOT.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="GameOutlet-jvm-1.0-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/build/libs</output-path>
+    <root id="archive" name="GameOutlet-jvm-1.0-SNAPSHOT.jar" />
+  </artifact>
+</component>
