--- .idea/artifacts/common_desktop_1_0_SNAPSHOT.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="common-desktop-1.0-SNAPSHOT">
-    <output-path>$PROJECT_DIR$/common/build/libs</output-path>
-    <root id="archive" name="common-desktop-1.0-SNAPSHOT.jar">
-      <element id="module-output" name="GameOutlet.common.desktopMain" />
-    </root>
-  </artifact>
-</component>
