--- .idea/artifacts/desktop_jvm_1_0_SNAPSHOT.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="desktop-jvm-1.0-SNAPSHOT">
-    <output-path>$PROJECT_DIR$/desktop/build/libs</output-path>
-    <root id="archive" name="desktop-jvm-1.0-SNAPSHOT.jar">
-      <element id="module-output" name="GameOutlet.desktop.jvmMain" />
-    </root>
-  </artifact>
-</component>
