--- android/src/main/java/appoutlet/android/MainActivity.kt
@@ -1,18 +1,0 @@
-package appoutlet.android
-
-import appoutlet.common.App
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.material.MaterialTheme
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            MaterialTheme {
-                App()
-            }
-        }
-    }
-}
