--- common/build.gradle.kts
@@ -1,71 +1,0 @@
-import org.jetbrains.compose.compose
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-    id("com.android.library")
-}
-
-group = "appoutlet"
-version = "1.0-SNAPSHOT"
-
-kotlin {
-    android()
-
-    jvm("desktop") {
-        jvmToolchain(11)
-    }
-
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(compose.runtime)
-                api(compose.foundation)
-                api(compose.material)
-            }
-        }
-
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-
-        val androidMain by getting {
-            dependencies {
-                api(libs.androidx.appcompat)
-                api(libs.androidx.core)
-            }
-        }
-
-        val androidTest by getting {
-            dependencies {
-                implementation("junit:junit:4.13.2")
-            }
-        }
-
-        val desktopMain by getting {
-            dependencies {
-                api(compose.preview)
-            }
-        }
-
-        val desktopTest by getting
-    }
-}
-
-android {
-    compileSdk = 33
-
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
-
-    defaultConfig {
-        minSdk = 24
-        targetSdk = 33
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-}
