--- desktop/build.gradle.kts
@@ -1,39 +1,0 @@
-import org.jetbrains.compose.compose
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
-
-plugins {
-    kotlin("multiplatform")
-    id("org.jetbrains.compose")
-}
-
-group = "appoutlet"
-version = "1.0-SNAPSHOT"
-
-
-kotlin {
-    jvm {
-        jvmToolchain(11)
-        withJava()
-    }
-
-    sourceSets {
-        val jvmMain by getting {
-            dependencies {
-                implementation(project(":common"))
-                implementation(compose.desktop.currentOs)
-            }
-        }
-        val jvmTest by getting
-    }
-}
-
-compose.desktop {
-    application {
-        mainClass = "MainKt"
-        nativeDistributions {
-            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
-            packageName = "GameOutlet"
-            packageVersion = "1.0.0"
-        }
-    }
-}
