--- desktop/src/jvmMain/kotlin/Main.kt
@@ -1,12 +1,0 @@
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-    Window(onCloseRequest = ::exitApplication) {
-        MaterialTheme {
-            Text("Hello World")
-        }
-    }
-}
