+++ gradle.properties
@@ -1,6 +1,4 @@
 kotlin.code.style=official
-kotlin.native.enableDependencyPropagation=false
-android.useAndroidX=true
 kotlin.version=1.7.20
 agp.version=7.3.0
 compose.version=1.2.2
