--- gradle/libs.versions.toml
@@ -1,7 +1,0 @@
-[plugins]
-
-
-[libraries]
-androidx-activityCompose = "androidx.activity:activity-compose:1.5.0"
-androidx-appcompat = "androidx.appcompat:appcompat:1.6.1"
-androidx-core = "androidx.core:core-ktx:1.9.0"
