+++ settings.gradle.kts
@@ -8,13 +8,9 @@ pluginManagement {
 
     plugins {
         kotlin("multiplatform").version(extra["kotlin.version"] as String)
-        kotlin("android").version(extra["kotlin.version"] as String)
-        id("com.android.application").version(extra["agp.version"] as String)
-        id("com.android.library").version(extra["agp.version"] as String)
         id("org.jetbrains.compose").version(extra["compose.version"] as String)
     }
 }
 
 rootProject.name = "GameOutlet"
 
-include(":android", ":desktop", ":common")
